feat: change the structure of classes for siamese networks
