+++ deeppavlov/configs/ranking/ranking_insurance.json
@@ -55,26 +55,23 @@
         "in": ["x_proc"],
         "in_y": ["y"],
         "out": ["y_predicted"],
-        "name": "siamese_model",
+        "name": "bilstm_nn",
         "device_num": 0,
         "train_now": true,
-        "network": {
-          "name": "bilstm_nn",
-          "len_vocab": "#siam_vocab.len",
-          "use_matrix": "#preproc.use_matrix",
-          "max_sequence_length": "#preproc.max_sequence_length",
-          "emb_matrix": "#embeddings.emb_mat",
-          "embedding_dim": "#siam_embedder.dim",
-          "seed": 243,
-          "reccurent": "bilstm",
-          "max_pooling": true,
-          "shared_weights": true,
-          "hidden_dim": 300,
-          "learning_rate": 1e-3,
-          "triplet_loss": true,
-          "hard_triplets": false,
-          "margin": 0.1
-        },
+        "len_vocab": "#siam_vocab.len",
+        "use_matrix": "#preproc.use_matrix",
+        "max_sequence_length": "#preproc.max_sequence_length",
+        "emb_matrix": "#embeddings.emb_mat",
+        "embedding_dim": "#siam_embedder.dim",
+        "seed": 243,
+        "reccurent": "bilstm",
+        "max_pooling": true,
+        "shared_weights": true,
+        "hidden_dim": 300,
+        "learning_rate": 1e-3,
+        "triplet_loss": true,
+        "hard_triplets": false,
+        "margin": 0.1,
         "batch_size": 256,
         "save_path": "insurance_ranking/model_weights.h5",
         "load_path": "insurance_ranking/model_weights.h5",
