+++ deeppavlov/configs/ranking/ranking_ubuntu_v2.json
@@ -53,21 +53,18 @@
         "in": ["x_proc"],
         "in_y": ["y"],
         "out": ["y_predicted"],
-        "name": "siamese_model",
+        "name": "bilstm_nn",
         "device_num": 0,
         "train_now": true,
-        "network": {
-          "name": "bilstm_nn",
-          "len_vocab": "#siam_vocab.len",
-          "use_matrix": "#preproc.use_matrix",
-          "max_sequence_length": "#preproc.max_sequence_length",
-          "embedding_dim": "#siam_embedder.dim",
-          "seed": 243,
-          "hidden_dim": 300,
-          "emb_matrix": "#embeddings.emb_mat",
-          "learning_rate": 1e-3,
-          "triplet_loss": false
-        },
+        "len_vocab": "#siam_vocab.len",
+        "use_matrix": "#preproc.use_matrix",
+        "max_sequence_length": "#preproc.max_sequence_length",
+        "embedding_dim": "#siam_embedder.dim",
+        "seed": 243,
+        "hidden_dim": 300,
+        "emb_matrix": "#embeddings.emb_mat",
+        "learning_rate": 1e-3,
+        "triplet_loss": false,
         "batch_size": 256,
         "interact_pred_num": 3,
         "save_path": "ubuntu_v2_ranking/model_weights.h5",
