+++ deeppavlov/configs/ranking/ranking_ubuntu_v2_memnet.json
@@ -52,17 +52,14 @@
         "in": ["x_proc"],
         "in_y": ["y"],
         "out": ["y_predicted"],
-        "name": "siamese_model",
+        "name": "qa_memnet",
         "device_num": 0,
         "train_now": true,
-        "network": {
-          "name": "qa_memnet",
-          "voc_siz": "#siam_vocab.len",
-          "seq_len": "#preproc.max_sequence_length",
-          "emb_dim": "#siam_embedder.dim",
-          "emb_matrix": "#embeddings.emb_mat",
-          "lstm_dim": 300
-        },
+        "voc_siz": "#siam_vocab.len",
+        "seq_len": "#preproc.max_sequence_length",
+        "emb_dim": "#siam_embedder.dim",
+        "emb_matrix": "#embeddings.emb_mat",
+        "lstm_dim": 300,
         "batch_size": 256,
         "num_context_turns": "#preproc.num_context_turns",
         "save_path": "ubuntu_v2_memnet_ranking/model_weights.h5",
