+++ deeppavlov/configs/ranking/ranking_ubuntu_v2_mt.json
@@ -54,24 +54,20 @@
         "in": ["x_proc"],
         "in_y": ["y"],
         "out": ["y_predicted"],
-        "name": "siamese_model",
+        "name": "bilstm_gru_nn",
         "device_num": 0,
         "train_now": true,
-        "network": {
-          "name": "bilstm_gru_nn",
-          "use_matrix": "#preproc.use_matrix",
-          "num_context_turns": "#preproc.num_context_turns",
-          "len_vocab": "#siam_vocab.len",
-          "max_sequence_length": "#preproc.max_sequence_length",
-          "embedding_dim": "#siam_embedder.dim",
-          "emb_matrix": "#embeddings.emb_mat",
-          "seed": 243,
-          "hidden_dim": 300,
-          "learning_rate": 1e-3,
-          "triplet_loss": false
-        },
-        "batch_size": 256,
+        "use_matrix": "#preproc.use_matrix",
         "num_context_turns": "#preproc.num_context_turns",
+        "len_vocab": "#siam_vocab.len",
+        "max_sequence_length": "#preproc.max_sequence_length",
+        "embedding_dim": "#siam_embedder.dim",
+        "emb_matrix": "#embeddings.emb_mat",
+        "seed": 243,
+        "hidden_dim": 300,
+        "learning_rate": 1e-3,
+        "triplet_loss": false,
+        "batch_size": 256,
         "save_path": "ubuntu_v2_mt_ranking/model_weights.h5",
         "load_path": "ubuntu_v2_mt_ranking/model_weights.h5"
       }
