--- deeppavlov/models/ranking/siamese_network.py
@@ -1,7 +1,0 @@
-from abc import abstractmethod
-
-class SiameseNetwork():
-
-    @abstractmethod
-    def load_initial_emb_matrix(self):
-        pass
